requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
boto3==1.24.79
botocore==1.27.79
certifi==2021.10.8
cffi==1.15.1
distlib==0.3.5
Django==4.1
django-storages==1.13.1
djangorestframework==3.13.1
filelock==3.7.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
jmespath==1.0.1
outcome==1.2.0
Pillow==9.2.0
platformdirs==2.5.2
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
s3transfer==0.6.0

six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.11
virtualenv==20.16.2
wsproto==1.1.0